hardware/envSensorService/requirements.txt
Adafruit_DHT==1.4.0
smbus2==0.4.3
websockets==12.0
pyserial==3.5
